Fix all installation scripts. Powershell script does all, but Linux user-centric install scripts can all be executed independently if required.

+++ scripts/install/installChrome.sh
@@ -2,6 +2,10 @@
 
 set -euo pipefail
 
-curl -fSL https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -o /tmp/google-chrome-stable_current_amd64.deb
-sudo dpkg -i /tmp/google-chrome-stable_current_amd64.deb
-rm /tmp/google-chrome-stable_current_amd64.deb
+sudo apt install -y fonts-liberation xdg-utils
+
+if [[ $(which google-chrome | wc -l) == 0 ]]; then
+    curl -fSL https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -o /tmp/google-chrome-stable_current_amd64.deb
+    sudo dpkg -i /tmp/google-chrome-stable_current_amd64.deb
+    rm /tmp/google-chrome-stable_current_amd64.deb
+fi
